+++ scripts/install-plugin.sh
@@ -46,8 +46,8 @@ then
     nvim --headless +"CocInstall -sync coc-word|qall"
     nvim --headless +"CocInstall -sync coc-dictionary|qall"
     nvim --headless +"CocInstall -sync coc-webview|qall"
-    #nvim --headless +"CocInstall -sync coc-markdown-preview-enhanced|qall"
+    nvim --headless +"CocInstall -sync coc-markdown-preview-enhanced|qall"
     #nvim --headless +"CocInstall -sync coc-ccls|qall"
-    #nvim --headless +"CocInstall -sync coc-explorer|qall"
+    nvim --headless +"CocInstall -sync coc-explorer|qall"
 fi